backend/requirements.txt
openai==1.58.1
fastapi==0.115.4
uvicorn==0.32.0
python-multipart==0.0.17
sqlalchemy==2.0.36
psycopg2-binary
alembic==1.14.0
uvicorn[standard]
python-jose[cryptography]==3.3.0
stripe==11.3.0
pydantic[email]==2.9.2
httpx==0.27.2
sse-starlette==2.1.3
markdown-to-json==2.1.2
python-dateutil==2.9.0.post0
fastapi_cache2==0.2.2
postmarker==1.0
